Update walkA0.py
Added ability to have more time points of integration than data points. This is done by having a data file with real data points every m entries and dummy values in the other ones. This should only be done if it is also done in the optimization procedure prior to this. Otherwise the scaling will be off by this factor.